diagonal

+++ pointers_arrays_strings/8-print_diagsums.c
@@ -8,7 +8,7 @@
 void print_diagsums(int *a, int size)
 {
 int i;
-unsigned int sum1 = 0, sum2 = 0;
+unsigned long sum1 = 0, sum2 = 0;
 for (i = 0; i < size; i++)
 {
 sum1 = sum1 + a[i * size + i];
